requirements.txt
Flask==2.0.1
requests==2.26.0
numpy==1.21.2
google_cloud_secret_manager==2.7.1
pyyaml
-i https://google-coral.github.io/py-repo/
tflite_runtime==2.5.0